myproject/requirements.txt
Django>=4.1,<4.2
psycopg2-binary>=2.9,<3.0
django-environ>=0.7,<1.0
pika==1.3.1
djangorestframework
django-environ==0.11.2